requirements.txt
ruamel.yaml
tenacity~=9.0.0
requests~=2.32.3
pydantic~=2.9.2
